requirements.txt
cycler==0.10.0
Django==2.2.3
django-bootstrap-datepicker-plus==3.0.5
django-datetimepicker==3.14
django-multiselectfield==0.1.8
django-widget-tweaks==1.4.5
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
pandas==0.25.1
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyparsing==2.4.2
pyserial==3.4
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
South==1.0.2
sqlparse==0.3.0
